% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGMQLScalaLib.R
\docType{package}
\name{RGMQLScalaLib}
\alias{RGMQLScalaLib}
\alias{RGMQLScalaLib-package}
\alias{RGMQLScalaLib-package}
\title{RGMQL scala API libraries}
\description{
RGMQLScalaLib
}
\details{
Is a data package containing compiled libraries in jar format used to
invoke GMQL scala API

This library is used in the package RGQML
}
