% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGMQLScalaLibs.R
\docType{package}
\name{RGMQLScalaLibs}
\alias{RGMQLScalaLibs}
\alias{RGMQLScalaLibs-package}
\alias{RGMQLScalaLibs-package}
\title{RGMQL scala API libraries}
\description{
RGMQLScalaLibs
}
\details{
Is a data package containing two compiled libraries in jar format used to
invoke GMQL scala API

This library is used in the package RGQML
}
