% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGMQLlib.R
\docType{package}
\name{RGMQLlib}
\alias{RGMQLlib}
\alias{RGMQLlib-package}
\alias{RGMQLlib-package}
\title{RGMQL scala API libraries}
\description{
RGMQLlib
}
\details{
Is a data package containing compiled libraries in jar format used to
invoke GMQL scala API

This library is used in the package RGQML
}
